requirements.txt
numpy
matplotlib
pillow
jupyter
torch
torchvision
streamlit